[CAY-1158] Replace the data type for linear algebra to float (#1159)
* Change Double to Float in Vectors

* Complete Breeze except some methods in MatrixOps

* Complete double -> float in breeze wrapper

* Float to float

* Float to float

* Minor logging (may be deleted)

* Minor fix

* Fix type issues

* minor

+++ dolphin/async/src/main/java/edu/snu/cay/dolphin/async/mlapps/gbt/GBTParameters.java
@@ -28,15 +28,15 @@ final class GBTParameters {
   }
 
   @NamedParameter(doc = "initial value of the step size", short_name = "init_step_size")
-  static final class StepSize implements Name<Double> {
+  static final class StepSize implements Name<Float> {
   }
 
   @NamedParameter(doc = "regularization constant for leaf", short_name = "lambda")
-  static final class Lambda implements Name<Double> {
+  static final class Lambda implements Name<Float> {
   }
 
   @NamedParameter(doc = "regularization constant for tree", short_name = "gamma")
-  static final class Gamma implements Name<Double> {
+  static final class Gamma implements Name<Float> {
   }
 
   @NamedParameter(doc = "maximum depth of model trees", short_name = "max_depth_of_tree")
